Moved Course Selection Validator
